[refactor] 制作フェーズ用のコントローラーをCreatePhaseディレクトリに移動
