+++ Assets/S2SData/CreateToInvasionData.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 81c3f1502d715b44ea6070d3f59b5678
-folderAsset: yes
-DefaultImporter:
+guid: 66aec002291057341bcb6f1da637634b
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
