+++ Assets/Scenes/CreatePhase.unity
@@ -443,6 +443,7 @@ MonoBehaviour:
   mazeData: {fileID: 11400000, guid: a76e0ed324fb0cf4ea3980131a4f296f, type: 2}
   tile: {fileID: 2227749310468335157, guid: 26a8b4975ac1871409a5e906e043cdeb, type: 3}
   reRollButton: {fileID: 777972593}
+  createToInvasionData: {fileID: 11400000, guid: 66aec002291057341bcb6f1da637634b, type: 2}
 --- !u!1 &348158059
 GameObject:
   m_ObjectHideFlags: 0
@@ -952,7 +953,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d2025fe8d6399ff47819f55b35d4f77c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  mazeController: {fileID: 321828198}
+  mazeCreationController: {fileID: 321828198}
 --- !u!1 &980838299
 GameObject:
   m_ObjectHideFlags: 0
