+++ Assets/Scripts/CreatePhase/CreationSceneController.cs
@@ -1,0 +1,19 @@
+using UnityEngine;
+
+namespace CreatePhase
+{
+    public class CreationSceneController : MonoBehaviour
+    {
+        // Start is called before the first frame update
+        void Start()
+        {
+        
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+        
+        }
+    }
+}
